Prune obsolete session state keys
